Implement a Colour FarmJob
All that it adds is `show_lut` and `shot_lut`

The DistantColourQuicktimeFarmJob will find the colour files if you don't let us know what they are.

It does this with a new submodule in the distant_vfx python library.

distant_vfx.colour
It's not the most simple way to get the correct colour file...

It will find the shot/plates folder, group the folders in there like this:
```python
{
  <descriptor>: {
    <version>: {
      "project": "dst",
      "sequence": <sequence>,
      "shot": <shot>,
      "descriptor": <descriptor>,
      "version": <version>,
      "filename": <foldername>,
    }
  }
}
```

It does this and then searches through the "bg" plates, highest version to lowest looking for a cube file.
If it cannot find any in "bg" it will search the rest of the descriptors.
If it still cannot find any, it will then search the "post/scan" directory because that was where I previously was putting the CDL files.
Woops...Whatever, legacy bug.